refactor: splitting Resource Providers out to their own package
This is required to unblock #7951, which has a circular reference otherwise
